Clarify Loop,Count param
The phrase "can be an expression" was used throughout the v1 docs, and implied "but may not be an expression". That implication no longer holds true in v2. This code works fine on v1 but not on v2, and hopefully this change clarifies it a bit:

```autohotkey
var := 2
Loop, %var%
   MsgBox, %A_Index%
```