Functions: Clarify that = and := function identically
On the page for *Variables and Expressions*, the `=` operator is described as follows.
"The traditional method uses the equal sign operator (=) to assign **unquoted** literal strings or variables enclosed in percent signs... By contrast, the expression method uses the colon-equal operator (:=) to store numbers, **quoted** strings, and other types of expressions."

To prevent confusion, this note clarifies that the two operators function identically when used to define optional parameters.