Add example frames
Mainly for adding the possibility to get a direct link to the example and providing a description in a more readable way. Previously this was achieved with overusing headings in some example sections. Every example has a descriptive anchor now. Code boxes in such frames doesn't need the NoIndent class. Every Example heading was renamed to Examples for consistency.