Add interaction with changing sendmode and kN
Added a note on interaction between SendMode and the key-delay at the same time. 

As per the example, `#Hotstring SE k10` will change the send mode, but the key-delay change will be ignored if the current sendmode is SendInput. Therefore, it's necessary to use two lines:

```
#Hotstring SE
#Hotstring k10
```

Probably should fix this as an issue, but settling for documenting it for now.