Add workaround for broken CHM
There appears to be a bug in HHC that causes it to produce a broken CHM,
when all source files are valid.  It's purportedly due to an error in the
compression module of HHC, and a workaround is to add a blank file to the
end of the file list.  Using NUL saves actually having a blank file in
the project directory.