Add @Ahk2Exe-Set example
The example changes the compiled script's displayed name in Task Manager.

From https://learn.microsoft.com/en-us/windows/win32/menurc/versioninfo-resource:

FileDescription - File description to be presented to users. This string may be displayed in a list box when the user is choosing files to install—for example, Keyboard Driver for AT-Style Keyboards. This string is required.